Remove controller and failing test
